Scalar Measures of Fit for Binary Variable Models

Calculates scalar measures of fit for models with binary dependent variables along the lines described in Long (1997) and Long and Freese (2005).

binfit(mod)

Arguments

mod
A model of class glm with family=binomial.

Value

A named vector of scalar measures of fit

Details

binfit calculates scalar measures of fit (many of which are pseudo-R-squared measures) to describe how well a model fits data with a binary dependent variable.

References

Long, J.S. 1997. Regression Models for Categorical and Limited Dependent Variables. Thousand Oaks, CA: Sage.

Long, J.S. and J. Freese. 2005. Regression Models for Categorical Outcomes Using Stata, 2nd ed. College Station, TX: Stata Press.

Examples

data(france)
left.mod <- glm(voteleft ~ male + age + retnat +
        poly(lrself, 2), data=france, family=binomial)
binfit(left.mod)

#>                    Names1    vals1                       Names2    vals2
#> 1 Log-Lik Intercept Only: -363.176          Log-Lik Full Model: -155.570
#> 2                 D(535):  311.139                       LR(6):  415.213
#> 3                                                    Prob > LR:    0.000
#> 4          McFadden's R2:    0.572           McFadden's Adk R2:    0.552
#> 5      ML (Cox-Snell) R2:    0.535 Cragg-Uhler (Nagelkerke) R2:    0.725
#> 6  McKelvey & Zavoina R2:    0.791                  Efron's R2:    0.641
#> 7               Count R2:    0.882                Adj Count R2:    0.700
#> 8                    BIC:  355.206                         AIC:  325.139