Test for Combining Categories in Multinomial Logistic Regression Models.

Tests the null hypothesis that categories can be combined in Multinomial Logistic Regression Models

combTest(obj)

Arguments

obj
An object of class multinom.

Value

A matrix of test statistics and p-values.

Examples

library(nnet)
data(france)
mnl.mod <- multinom(vote ~ age + male + retnat + lrself, data=france)

#> # weights:  35 (24 variable)
#> initial  value 872.315349 
#> iter  10 value 655.272636
#> iter  20 value 559.902797
#> iter  30 value 551.176433
#> final  value 551.169697 
#> converged

combTest(mnl.mod)

#>           Estimate p-value
#> PCF-PS     20.280  0.001  
#> PCF-Green  33.661  0.000  
#> PCF-RPR   131.520  0.000  
#> PCF-UDF   113.322  0.000  
#> PS-Green   27.497  0.000  
#> PS-RPR    122.279  0.000  
#> PS-UDF    105.010  0.000  
#> Green-RPR  75.026  0.000  
#> Green-UDF  62.751  0.000  
#> RPR-UDF    18.913  0.002