Average Effects Plot for Multinomial Logistic Regression

Produces a plot of average effects for one variable while holding the others constant at observed values.

mnlAveEffPlot(obj, varname, data, R = 1500, nvals = 25, plot = TRUE, ...)

Arguments

obj
An object of class multinom.

varname
A string indicating the variable for which the plot is desired.

data
The data used to estimate obj.

R
Number of simulations used to generate confidence bounds.

nvals
Number of evaluation points for the predicted probabilities.

plot
Logical indicating whether a plot should be produced (if TRUE) or numerical results should be returned (if FALSE).

...
Other arguments to be passed down to xyplot.

Value

Either a plot or a data frame with variables

mean

The average effect (i.e., predicted probability)

lower

The lower 95% confidence bound

upper

The upper 95% confidence bound

y

The values of the dependent variable being predicted

x

The values of the independent variable being manipulated

References

Hanmer, M.J. and K.O. Kalkan. 2013. ‘Behind the Curve: Clarifying the Best Approach to Calculating Predicted Probabilities and Marginal Effects from Limited Dependent Variable Models’. American Journal of Political Science. 57(1): 263-277.

Examples

library(nnet)
data(france)
mnl.mod <- multinom(vote ~ age + male + retnat + lrself, data=france)

#> # weights:  35 (24 variable)
#> initial  value 872.315349 
#> iter  10 value 655.272636
#> iter  20 value 559.902797
#> iter  30 value 551.176433
#> final  value 551.169697 
#> converged

if (FALSE) mnlAveEffPlot(mnl.mod, "lrself", data=france)