Print Statistically Significant MNL Coefficients

By default, the summary for objects of class multinom is not particularly helpful. It still requires a lot of work on the part of the user to figure out which coefficients are significantly different from zero and which ones are not. mnlSig solves this problem by either flagging significant coefficients with an asterisk or only printing significant coefficients, leaving insignificant ones blank.

mnlSig(
  obj,
  pval = 0.05,
  two.sided = TRUE,
  flag.sig = TRUE,
  insig.blank = FALSE
)

Arguments

obj
A model object of class multinom.

pval
The desired Type I error rate to identify coefficients as statistically significant.

two.sided
Logical indicating whether calculated p-values should be two-sided (if TRUE) or one-sided (if FALSE).

flag.sig
Logical indicating whether an asterisk should be placed beside coefficients which are significant at the pval level.

insig.blank
Logical indicating whether coefficients which are not significant at the pval level should be blank in the output.

Value

A data frame suitable for printing with the (optionally significance-flagged) coefficients from a multinomial logit model.

Examples

library(nnet)
data(france)
mnl.mod <- multinom(vote ~ retnat + male + retnat + lrself, data=france)

#> # weights:  30 (20 variable)
#> initial  value 872.315349 
#> iter  10 value 626.581729
#> iter  20 value 566.113965
#> iter  30 value 562.683607
#> final  value 562.683588 
#> converged

mnlSig(mnl.mod)

#>       (Intercept) retnatSame retnatWorse    male lrself
#> PS         1.193     -1.281*     -0.988   0.206  0.381*
#> Green     -2.112*    -0.933      -0.461  -0.199  0.863*
#> RPR       -8.101*    -1.876*     -1.182   0.606  2.122*
#> UDF       -7.844*    -0.798      -0.227   1.115* 1.912*