Fit Statistics and Specification Test for Multinomial Logistic Regression

Provides fit statistics (pseudo R-squared values) and the Fagerland, Hosmer and Bonfi (2008) specification test for Multinomial Logistic Regression models.

mnlfit(obj, permute = FALSE)

Arguments

obj
An object of class multinom

permute
Logical indicating whether to check all base categories for the Fagerland et. al. specification test.

Value

A list with elements:

result

Fit statistics.

permres

The results of the base category permutation exercise.

References

Fagerland M. W., D. W. Hosmer and A. M. Bonfi. 2008. ‘Multinomial goodness-of-fit tests for logistic regression models.’ Statistics in Medicine. 27(21): 4238-4253.

Examples

library(nnet)
data(france)
mnl.mod <- multinom(vote ~ age + male + retnat + lrself, data=france)

#> # weights:  35 (24 variable)
#> initial  value 872.315349 
#> iter  10 value 655.272636
#> iter  20 value 559.902797
#> iter  30 value 551.176433
#> final  value 551.169697 
#> converged

mnlfit(mnl.mod)

#> Warning: 42% of expected counts < 5

#> $result
#>                               Estimate    p-value
#> Fagerland, Hosmer and Bonfi 45.0880444 0.06232676
#> Count R2                     0.5977860         NA
#> Count R2 (Adj)               0.3292308         NA
#> ML R2                        0.5990507         NA
#> McFadden R2                  0.3100393         NA
#> McFadden R2 (Adj)            0.2737367         NA
#> Cragg-Uhler(Nagelkerke) R2   0.6322150         NA
#> 
#> attr(,"class")
#> [1] "mnlfit"