Proportional and Expected Proportional Reductions in Error

Calculates proportional reduction in error (PRE) and expected proportional reduction in error (epre) from Herron (1999).

pre(mod1, mod2 = NULL, sim = FALSE, R = 2500)

Arguments

mod1
A model of class glm (with family binomial), polr or multinom for which (e)PRE will be calculated.

mod2
A model of the same class as mod1 against which proportional reduction in error will be measured. If NULL, the null model will be used.

sim
A logical argument indicating whether a parametric bootstrap should be used to calculate confidence bounds for (e)PRE. See Details for more information.

R
Number of bootstrap samples to be drawn if sim=TRUE.

Value

An object of class pre, which is a list with the following elements:

pre

The proportional reduction in error

epre

The expected proportional reduction in error

m1form

The formula for model 1

m2form

The formula for model 2

pcp

The percent correctly predicted by model 1

pmc

The percent correctly predicted by model 2

epcp

The expected percent correctly predicted by model 1

epmc

The expected percent correctly predicted by model 2

pre.sim

A vector of bootstrapped PRE values if sim=TRUE

epre.sim

A vector of bootstrapped ePRE values if sim=TRUE

Details

Proportional reduction in error is calculated as a function of correct and incorrect predictions (and the probabilities of correct and incorrect predictions for ePRE). When sim=TRUE, a parametric bootstrap will be used that draws from the multivariate normal distribution centered at the coefficient estimates from the model and using the estimated variance-covariance matrix of the estimators as Sigma. This matrix is used to form R versions of XB and predictions are made for each of the R different versions of XB. Confidence intervals can then be created from the bootstrap sampled (e)PRE values.

References

Herron, M. 1999. Postestimation Uncertainty in Limited Dependent Variable Models. Political Analysis 8(1): 83--98.

Examples

data(france)
left.mod <- glm(voteleft ~ male + age + retnat +
        poly(lrself, 2), data=france, family=binomial)
pre(left.mod)

#> mod1:  voteleft ~ male + age + retnat + poly(lrself, 2) 
#> mod2:  voteleft ~ 1 
#> 
#> Analytical Results
#>  PMC =  0.607 
#>  PCP =  0.882 
#>  PRE =  0.700 
#> ePMC =  0.523 
#> ePCP =  0.825 
#> ePRE =  0.634 
#>