Print Confidence Intervals for Predicted Probabilities and Their Differences

Print method for output from the probci function.

# S3 method for diffci
print(x, ..., digits = 4, filter = NULL, const = NULL, onlySig = FALSE)

Arguments

x
A object of class diffci produced by probci.

...
Other arguments to be passed down to print, currently unimplemented.

digits
How many digits to round output.

filter
A named list of values where the names indicate the variable to be filtered and the values in the vector indicate the values to include for the filtering variable.

const
A string identifying the name of the variable to be held constant across comparisons.

onlySig
Logical indicating whether all differes should be displayed or only those significant at the 95% two-tailed level.

Value

An data frame with the following variables:

variables

The variables and the values at which they are held constant. For example, tmp1 would be the first value of tmp used in the probability calculation and tmp2 would be the second value of tmp used in the probability calculation.

pred_prob

The difference in predicted probability given the following change in X: tmp2-tmp1.

lower, upper

The lower and upper 95% confidence bounds.

Examples

data(france)
left.mod <- glm(voteleft ~ male + age + retnat +
        poly(lrself, 2, raw=TRUE), data=france, family=binomial)
out <- probci(left.mod, france, numQuantVals=3,
    changeX=c("retnat", "lrself"))
print(out, filter=list(retnat=c("Better", "Worse")))

#>    retnat1 lrself1 retnat2 lrself2 pred_prob   lower   upper
#> 1   Better  1.0000   Worse  1.0000    0.0001 -0.0087  0.0139
#> 2   Better  1.0000  Better  5.5000   -0.6026 -0.7247 -0.4673
#> 3   Better  1.0000   Worse  5.5000   -0.5908 -0.7144 -0.4463
#> 4   Better  1.0000  Better 10.0000   -0.9939 -0.9986 -0.9515
#> 5   Better  1.0000   Worse 10.0000   -0.9939 -0.9987 -0.9512
#> 6    Worse  1.0000  Better  5.5000   -0.6038 -0.7266 -0.4652
#> 7    Worse  1.0000   Worse  5.5000   -0.5911 -0.7113 -0.4541
#> 8    Worse  1.0000  Better 10.0000   -0.9942 -0.9987 -0.9551
#> 9    Worse  1.0000   Worse 10.0000   -0.9942 -0.9986 -0.9545
#> 10  Better  5.5000   Worse  5.5000    0.0124 -0.1573  0.1763
#> 11  Better  5.5000  Better 10.0000   -0.3874 -0.5189 -0.2671
#> 12  Better  5.5000   Worse 10.0000   -0.3875 -0.5193 -0.2653
#> 13   Worse  5.5000  Better 10.0000   -0.3991 -0.5274 -0.2787
#> 14   Worse  5.5000   Worse 10.0000   -0.3988 -0.5264 -0.2794
#> 15  Better 10.0000   Worse 10.0000    0.0000 -0.0012  0.0020

print(out, filter=list(retnat=c("Better", "Worse")),
     const="lrself")

#>   retnat1 lrself1 retnat2 lrself2 pred_prob   lower  upper
#> 1  Better  1.0000   Worse  1.0000    0.0001 -0.0087 0.0139
#> 2  Better  5.5000   Worse  5.5000    0.0124 -0.1573 0.1763
#> 3  Better 10.0000   Worse 10.0000    0.0000 -0.0012 0.0020

print(out, const="retnat")

#>   retnat1 lrself1 retnat2 lrself2 pred_prob   lower   upper
#> 1  Better  1.0000  Better  5.5000   -0.6026 -0.7247 -0.4673
#> 2  Better  1.0000  Better 10.0000   -0.9939 -0.9986 -0.9515
#> 3    Same  1.0000    Same  5.5000   -0.5443 -0.6392 -0.4322
#> 4    Same  1.0000    Same 10.0000   -0.9949 -0.9988 -0.9623
#> 5   Worse  1.0000   Worse  5.5000   -0.5911 -0.7113 -0.4541
#> 6   Worse  1.0000   Worse 10.0000   -0.9942 -0.9986 -0.9545
#> 7  Better  5.5000  Better 10.0000   -0.3874 -0.5189 -0.2671
#> 8    Same  5.5000    Same 10.0000   -0.4480 -0.5515 -0.3527
#> 9   Worse  5.5000   Worse 10.0000   -0.3988 -0.5264 -0.2794